requirements.txt
Flask==2.0.1
Pillow==8.4.0
pytesseract==0.3.8
python-magic-bin==0.4.14
